+++ order/src/main/java/com/toanyone/order/application/dto/service/response/HubFindResponseDto.java
@@ -1,4 +1,4 @@
-package com.toanyone.order.application.dto;
+package com.toanyone.order.application.dto.service.response;
 
 import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
