+++ order/src/main/java/com/toanyone/order/application/service/HubService.java
@@ -1,6 +1,6 @@
 package com.toanyone.order.application.service;
 
-import com.toanyone.order.application.dto.HubFindResponseDto;
+import com.toanyone.order.application.dto.service.response.HubFindResponseDto;
 import com.toanyone.order.common.dto.SingleResponse;
 import org.springframework.http.ResponseEntity;
