+++ order/src/main/java/com/toanyone/order/application/service/OrderService.java
@@ -1,14 +1,14 @@
 package com.toanyone.order.application.service;
 
-import com.toanyone.order.application.dto.HubFindResponseDto;
-import com.toanyone.order.application.dto.SlackMessageRequestDto;
-import com.toanyone.order.application.dto.StoreFindResponseDto;
-import com.toanyone.order.application.dto.request.OrderCancelServiceDto;
-import com.toanyone.order.application.dto.request.OrderCreateServiceDto;
-import com.toanyone.order.application.dto.request.OrderFindAllCondition;
-import com.toanyone.order.application.dto.request.OrderSearchCondition;
-import com.toanyone.order.application.mapper.ItemRequestMapper;
-import com.toanyone.order.application.mapper.MessageConverter;
+import com.toanyone.order.application.dto.service.response.HubFindResponseDto;
+import com.toanyone.order.application.dto.service.response.StoreFindResponseDto;
+import com.toanyone.order.application.dto.service.request.OrderCancelServiceDto;
+import com.toanyone.order.application.dto.service.request.OrderCreateServiceDto;
+import com.toanyone.order.application.dto.service.request.OrderFindAllCondition;
+import com.toanyone.order.application.dto.service.request.OrderSearchCondition;
+import com.toanyone.order.infrastructure.kafka.OrderKafkaProducer;
+import com.toanyone.order.infrastructure.mapper.ItemRequestMapper;
+import com.toanyone.order.infrastructure.converter.OrderMessageConverter;
 import com.toanyone.order.common.dto.CursorPage;
 import com.toanyone.order.common.dto.SingleResponse;
 import com.toanyone.order.common.exception.OrderException;
@@ -48,8 +48,8 @@ public class OrderService {
     private final StoreService storeService;
     private final HubService hubService;
     private final ItemRequestMapper itemRequestMapper;
-    private final MessageConverter messageConverter;
-    private final OrderKafkaProducer orderKafkaProducer;
+    private final OrderMessageConverter orderMessageConverter;
+    private final OrderMessageProducer orderMessageProducer;
     private final RedisTemplate<String, Object> redisTemplate;
 
     @Transactional
@@ -87,12 +87,12 @@ public class OrderService {
 
         log.info("orderId: {}, userId: {}, totalPrice: {}", order.getId(), order.getUserId(), order.getTotalPrice());
 
-        DeliveryRequestMessage deliveryMessage = messageConverter.toOrderDeliveryMessage(request, order.getId(), receiveStore.getBody().getData().getHubId(), supplyStore.getBody().getData().getHubId());
+        DeliveryRequestMessage deliveryMessage = orderMessageConverter.toOrderDeliveryMessage(request, order.getId(), receiveStore.getBody().getData().getHubId(), supplyStore.getBody().getData().getHubId());
 
         redisTemplate.opsForValue().set(order.getId().toString(), deliveryMessage, Duration.ofMinutes(DELIVERY_REQUEST_EXPIRATION_MINUTES));
 
-        PaymentRequestMessage paymentMessage = messageConverter.toOrderPaymentMessage(order.getId(), order.getTotalPrice());
-        orderKafkaProducer.sendPaymentRequestMessage(paymentMessage, userId, role, slackId);
+        PaymentRequestMessage paymentMessage = orderMessageConverter.toOrderPaymentMessage(order.getId(), order.getTotalPrice());
+        orderMessageProducer.sendPaymentRequestMessage(paymentMessage, userId, role, slackId);
 
         return OrderCreateResponseDto.fromOrder(order);
     }
@@ -181,7 +181,7 @@ public class OrderService {
     private void cancelOrderAndSendPaymentCancelMessage(Order order, Long userId, String role, String slackId) {
         order.paymentCancelRequested();
         PaymentCancelMessage paymentMessage = PaymentCancelMessage.builder().orderId(order.getId()).build();
-        orderKafkaProducer.sendPaymentCancelMessage(paymentMessage, userId, role, slackId);
+        orderMessageProducer.sendPaymentCancelMessage(paymentMessage, userId, role, slackId);
     }
