+++ order/src/main/java/com/toanyone/order/application/service/StoreService.java
@@ -1,6 +1,6 @@
 package com.toanyone.order.application.service;
 
-import com.toanyone.order.application.dto.StoreFindResponseDto;
+import com.toanyone.order.application.dto.service.response.StoreFindResponseDto;
 import com.toanyone.order.common.dto.SingleResponse;
 import org.springframework.http.ResponseEntity;
