+++ order/src/main/java/com/toanyone/order/domain/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.toanyone.order.domain.repository;
 
-import com.toanyone.order.application.dto.request.OrderSearchCondition;
+import com.toanyone.order.application.dto.service.request.OrderSearchCondition;
 import com.toanyone.order.common.dto.CursorPage;
 import com.toanyone.order.domain.model.Order;
 import com.toanyone.order.infrastructure.repository.OrderQueryDslRepository;
