+++ order/src/main/java/com/toanyone/order/infrastructure/repository/OrderQueryDslRepository.java
@@ -1,7 +1,7 @@
 package com.toanyone.order.infrastructure.repository;
 
-import com.toanyone.order.application.dto.request.OrderFindAllCondition;
-import com.toanyone.order.application.dto.request.OrderSearchCondition;
+import com.toanyone.order.application.dto.service.request.OrderFindAllCondition;
+import com.toanyone.order.application.dto.service.request.OrderSearchCondition;
 import com.toanyone.order.common.dto.CursorPage;
 import com.toanyone.order.domain.model.Order;
