+++ order/src/main/java/com/toanyone/order/infrastructure/repository/OrderQueryDslRepositoryImpl.java
@@ -3,13 +3,12 @@ package com.toanyone.order.infrastructure.repository;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.toanyone.order.application.dto.request.OrderFindAllCondition;
-import com.toanyone.order.application.dto.request.OrderSearchCondition;
+import com.toanyone.order.application.dto.service.request.OrderFindAllCondition;
+import com.toanyone.order.application.dto.service.request.OrderSearchCondition;
 import com.toanyone.order.common.dto.CursorInfo;
 import com.toanyone.order.common.dto.CursorPage;
 import com.toanyone.order.domain.model.Order;
 import com.toanyone.order.common.dto.SortType;
-import com.toanyone.order.infrastructure.repository.OrderQueryDslRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.Hibernate;
