+++ order/src/main/java/com/toanyone/order/message/DeliveryRequestMessage.java
@@ -1,13 +1,11 @@
 package com.toanyone.order.message;
 
-import com.toanyone.order.application.dto.request.OrderCreateServiceDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Getter
 @Builder
